feat(Cell): implement new
This new function takes 4 booleans which represent the existence of a
wall in each direction of the cell and creates an array from it. It has
normalizes the diagonal by always having a line going from left to right
and from down to up, which makes the side that does not appear
predictable.