feat(maze_hex): implement break walls
This function is way bigger than its counterpart on maze, because for
the hexagonal maze each odd and even columns have different ways of
moving.
So, for instance, to go up and right on an even column, you need to add
+1 to x and -1 to y, while in an odd column, you would need to add +1 to
x and nothing to y.
This creates a need for a lot of different moving options that all have
different walls to break.